+++ src/adapters/web_socket.rs
@@ -61,10 +61,10 @@ impl Resource for RemoteResource {
 impl Remote for RemoteResource {
     fn connect(remote_addr: RemoteAddr) -> io::Result<ConnectionInfo<Self>> {
         let (peer_addr, url) = match remote_addr {
-            RemoteAddr::SocketAddr(addr) => {
+            RemoteAddr::Socket(addr) => {
                 (addr, Url::parse(&format!("ws://{}/message-io-default", addr)).unwrap())
             }
-            RemoteAddr::Path(path) => {
+            RemoteAddr::Str(path) => {
                 let url = Url::parse(&path).expect("A valid URL");
                 let addr = url
                     .socket_addrs(|| match url.scheme() {