+++ src/remote_addr.rs
@@ -9,36 +9,36 @@ use std::io::{self};
 /// It is usually used in [`crate::network::Network::connect()`] to specify the remote address.
 #[derive(Serialize, Deserialize, Clone, PartialEq, Eq, Hash, Debug)]
 pub enum RemoteAddr {
-    SocketAddr(SocketAddr),
-    Path(String),
+    Socket(SocketAddr),
+    Str(String),
 }
 
 impl RemoteAddr {
     /// Check if the `RemoteAddr` is a [`SocketAddr`].
     pub fn is_socket_addr(&self) -> bool {
-        matches!(self, RemoteAddr::SocketAddr(_))
+        matches!(self, RemoteAddr::Socket(_))
     }
 
-    /// Check if the `RemoteAddr` is a path.
-    pub fn is_path(&self) -> bool {
-        matches!(self, RemoteAddr::SocketAddr(_))
+    /// Check if the `RemoteAddr` is a string.
+    pub fn is_string(&self) -> bool {
+        matches!(self, RemoteAddr::Socket(_))
     }
 
     /// Extract the [`SocketAddr`].
     /// This function panics if the `RemoteAddr` do not represent a `SocketAddr`.
     pub fn socket_addr(&self) -> &SocketAddr {
         match self {
-            RemoteAddr::SocketAddr(addr) => addr,
+            RemoteAddr::Socket(addr) => addr,
             _ => panic!("The RemoteAddr must be a SocketAddr"),
         }
     }
 
-    /// Extract the path.
-    /// This function panics if the `RemoteAddr` do not represent a path.
-    pub fn path(&self) -> &str {
+    /// Extract the string.
+    /// This function panics if the `RemoteAddr` is not a `Str` variant.
+    pub fn string(&self) -> &str {
         match self {
-            RemoteAddr::Path(addr) => addr,
-            _ => panic!("The RemoteAddr must be a path"),
+            RemoteAddr::Str(addr) => addr,
+            _ => panic!("The RemoteAddr must be a String"),
         }
     }
 }
@@ -47,8 +47,8 @@ impl ToSocketAddrs for RemoteAddr {
     type Iter = std::option::IntoIter<SocketAddr>;
     fn to_socket_addrs(&self) -> io::Result<Self::Iter> {
         match self {
-            RemoteAddr::SocketAddr(addr) => addr.to_socket_addrs(),
-            RemoteAddr::Path(_) => Err(io::Error::new(
+            RemoteAddr::Socket(addr) => addr.to_socket_addrs(),
+            RemoteAddr::Str(_) => Err(io::Error::new(
                 io::ErrorKind::InvalidInput,
                 "The RemoteAddr is not a SocketAddr",
             )),
@@ -59,8 +59,8 @@ impl ToSocketAddrs for RemoteAddr {
 impl std::fmt::Display for RemoteAddr {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            RemoteAddr::SocketAddr(addr) => write!(f, "{}", addr),
-            RemoteAddr::Path(path) => write!(f, "{}", path),
+            RemoteAddr::Socket(addr) => write!(f, "{}", addr),
+            RemoteAddr::Str(string) => write!(f, "{}", string),
         }
     }
 }
@@ -69,7 +69,7 @@ impl std::fmt::Display for RemoteAddr {
 /// Instead of `ToSocketAddrs` that only can accept valid 'ip:port' string format,
 /// `ToRemoteAddr` accept any string without panic.
 /// If the string has the 'ip:port' format, it will be interpreted as a [`SocketAddr`],
-/// if not, it will be interpreted as a string path.
+/// if not, it will be interpreted as a string.
 pub trait ToRemoteAddr {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr>;
 }
@@ -77,8 +77,8 @@ pub trait ToRemoteAddr {
 impl ToRemoteAddr for &str {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
         Ok(match self.parse() {
-            Ok(addr) => RemoteAddr::SocketAddr(addr),
-            Err(_) => RemoteAddr::Path(self.to_string()),
+            Ok(addr) => RemoteAddr::Socket(addr),
+            Err(_) => RemoteAddr::Str(self.to_string()),
         })
     }
 }
@@ -97,19 +97,19 @@ impl ToRemoteAddr for &String {
 
 impl ToRemoteAddr for SocketAddr {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
-        Ok(RemoteAddr::SocketAddr(*self))
+        Ok(RemoteAddr::Socket(*self))
     }
 }
 
 impl ToRemoteAddr for SocketAddrV4 {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
-        Ok(RemoteAddr::SocketAddr(SocketAddr::V4(*self)))
+        Ok(RemoteAddr::Socket(SocketAddr::V4(*self)))
     }
 }
 
 impl ToRemoteAddr for SocketAddrV6 {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
-        Ok(RemoteAddr::SocketAddr(SocketAddr::V6(*self)))
+        Ok(RemoteAddr::Socket(SocketAddr::V6(*self)))
     }
 }
 
@@ -121,31 +121,31 @@ impl ToRemoteAddr for RemoteAddr {
 
 impl ToRemoteAddr for (&str, u16) {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
-        Ok(RemoteAddr::SocketAddr(self.to_socket_addrs().unwrap().next().unwrap()))
+        Ok(RemoteAddr::Socket(self.to_socket_addrs().unwrap().next().unwrap()))
     }
 }
 
 impl ToRemoteAddr for (String, u16) {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
-        Ok(RemoteAddr::SocketAddr(self.to_socket_addrs().unwrap().next().unwrap()))
+        Ok(RemoteAddr::Socket(self.to_socket_addrs().unwrap().next().unwrap()))
     }
 }
 
 impl ToRemoteAddr for (IpAddr, u16) {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
-        Ok(RemoteAddr::SocketAddr(self.to_socket_addrs().unwrap().next().unwrap()))
+        Ok(RemoteAddr::Socket(self.to_socket_addrs().unwrap().next().unwrap()))
     }
 }
 
 impl ToRemoteAddr for (Ipv4Addr, u16) {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
-        Ok(RemoteAddr::SocketAddr(self.to_socket_addrs().unwrap().next().unwrap()))
+        Ok(RemoteAddr::Socket(self.to_socket_addrs().unwrap().next().unwrap()))
     }
 }
 
 impl ToRemoteAddr for (Ipv6Addr, u16) {
     fn to_remote_addr(&self) -> io::Result<RemoteAddr> {
-        Ok(RemoteAddr::SocketAddr(self.to_socket_addrs().unwrap().next().unwrap()))
+        Ok(RemoteAddr::Socket(self.to_socket_addrs().unwrap().next().unwrap()))
     }
 }
 
@@ -155,15 +155,15 @@ mod tests {
     use std::net::{IpAddr, Ipv4Addr};
 
     #[test]
-    fn str_to_path() {
-        let path = "ws://domain:1234/socket";
-        assert_eq!(path, path.to_remote_addr().unwrap().path());
+    fn str_to_string() {
+        let string = "ws://domain:1234/socket";
+        assert_eq!(string, string.to_remote_addr().unwrap().string());
     }
 
     #[test]
-    fn string_to_path() {
-        let path = String::from("ws://domain:1234/socket");
-        assert_eq!(&path, path.to_remote_addr().unwrap().path());
+    fn string_to_string() {
+        let string = String::from("ws://domain:1234/socket");
+        assert_eq!(&string, string.to_remote_addr().unwrap().string());
     }
 
     #[test]